New MergeWithTagStructure test focusing on the stack overflow question 72922255
When merging pdfs on certain order getting "Tag structure copying failed"
https://stackoverflow.com/questions/72922255/when-merging-pdfs-on-certain-order-getting-tag-structure-copying-failed
10840.pdf, invoice.pdf, B.Tech-Cyber-180-Credits.docx.pdf
https://drive.google.com/file/d/1ELt-bPKFNJgUzvuu55c8sjUT3hI8fios/view?usp=sharing

testMergeLikeDeepanshuGood - This is the OP's code applied to his good case.

testMergeLikeDeepanshuBad - This is the OP's code applied to his bad case.

testMergeLikeDeepanshuButUnTaggedGood - This is the OP's code modified to merge without tagging applied to his good case.

testMergeLikeDeepanshuButUnTaggedBad - This is the OP's code modified to merge without tagging applied to his bad case.