New ParagraphContent test focusing on the stack overflow question 72929075
How to view the text of Itext7 com.itextpdf.layout.element.Paragraph in Java
https://stackoverflow.com/questions/72929075/how-to-view-the-text-of-itext7-com-itextpdf-layout-element-paragraph-in-java

testKishoresExample - This test shows how to extract the content from the Paragraph element given as example by the OP.

testSubParagraph - This test shows how to extract the content from a Paragraph element with a sub-paragraph.

testTable - This test shows how to extract the content from a Table element.