New VerifySignature test focusing on the stack overflow question 75005703
Adobe Acrobat Reader and Foxit reader show PDF Signature Valid but iText7 says otherwise
https://stackoverflow.com/questions/75005703/adobe-acrobat-reader-and-foxit-reader-show-pdf-signature-valid-but-itext7-says-o
3-HD-LAO-DONG-17.50.25-29.03.2022 (3).pdf
https://drive.google.com/file/d/1bvk5EckUrJ-dzehhQJyBGkM4QLrIF3Uk/view?usp=sharing

testVerify3HdLaoDong17_50_25_29_03_2022 - Indeed, the second signature fails to validate in iText. As it turns out, the CMS container of that signature has an eContent element in its EncapsulatedContentInfo, and iText uses it in its validation code. On the other hand, this element contains an empty octet string which makes no sense at all in a PDF signature, so one can also understand that lax validators ignore it.